fix: モードが切り替わらない不具合を修正
