+++ src/hooks/ThemeContext.jsx
@@ -1,0 +1,34 @@
+// import { useEffect } from "react";
+// import { useLocalStorage } from "react-use";
+
+import { createContext, useEffect } from "react";
+import { useLocalStorage } from "react-use";
+
+const initialTheme = () => {
+	if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
+		return "dark";
+	}
+	return "light";
+};
+
+const ThemeContext = createContext();
+
+const ThemeProvider = ({ children }) => {
+	const [theme, setTheme] = useLocalStorage("theme", initialTheme);
+
+	useEffect(() => {
+		document.documentElement.setAttribute("data-theme", theme);
+	}, [theme]);
+
+	const toggleTheme = () => {
+		setTheme(theme === "light" ? "dark" : "light");
+	};
+
+	return (
+		<ThemeContext.Provider value={{ theme, toggleTheme }}>
+			{children}
+		</ThemeContext.Provider>
+	);
+};
+
+export { ThemeProvider, ThemeContext };
