+++ src/main.jsx
@@ -2,9 +2,12 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
+import { ThemeProvider } from "./hooks/ThemeContext.jsx";
 
 createRoot(document.getElementById("root")).render(
 	<StrictMode>
-		<App />
+		<ThemeProvider>
+			<App />
+		</ThemeProvider>
 	</StrictMode>,
 );
